requirements.txt
opencv-python==4.8.1.78
torchvision==0.16.1
Flask==3.0.0
Pillow==10.1.0
pickle-mixin==1.0.2